Transport Layer: low-level read-line with new Buffer class
If the input stream gets a reader, the reader cannot be dropped, thus
it contains the next packet in certain cases which should not be happened.
So, the getLine() is reimplemented in low-level byte form, which assumes
that the line ends with NL (0x0a), and preferably with CR LF according
to RFC 2453 Transport Layer Protocol.

The version string reader method is rewritten accordingly.