complete a naive implementation of KEX calculation
Originally the host key was ignored, but otherwise the previously
implemented parts may work well in a slow, inefficient way.
The new code, the calculation is way too naive, as the different
algorithms are related to each other, so a host key algorithm may
affect an encryption algorithm just as a random example.
This correlation is not added, so I would say it is completely wrong,
but mostly it can work.

I add this to the repository as a commit for historical reasons,
I mean later it can be seen how not to implement the calculation.

Note that as of now the code does not send any SSH_MSG_KEXINIT. Yet.