kex: implement algo selection with placeholder KexAlgo class
Add KexAlgo class to encrypt / decrypt bytes in the future, but currently the methods
are dummy. It is configured by cipher and possibly mac in the future, again.

For the SSH_MSG_KEXINIT the algorithm selection is implemented with a limited number
of encyption (cipher) and mac algorithms. The compression is not supported
(always: "none").

The code is based on OpenSSH's implementation.